Adding a README file and environmental variables.
Now if the user wants to run the project with a Mongo Atlas Cloud
database it is only required to register the appropriate url in a .env
file at the root dir. If none .env file detected, the project will use
a local Mongo database instead.

deleteProduct function was removed from an inline onclick listener at
the Html file to the javascript.

*Minor bug fixed. To get product list when there are not products do
not launch error anymore.