Adding The Project Description
The project description and some instructions in the landing page were
added.
The login controller and login view were modified to show the reset
password url in case of request.
The user session was modified to expire after 5 hours of being created.
The admin-product-list was modified to add a little note for the user.
The add-product page had a minor UI bug with the logout hover, this was
solved by importing form.css before main.css which cause the head file
to be updated.
The cart title was also enhaced.